ml_project/requirements.txt
# Core ML libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Data processing
joblib>=1.3.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.23.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Code quality
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0

# Optional: Deep Learning (uncomment if needed)
# torch>=2.0.0
# tensorflow>=2.12.0

# Optional: Advanced ML (uncomment if needed)
# xgboost>=1.7.0
# lightgbm>=3.3.0
# catboost>=1.2.0
